requirements.txt
torch>=2.7.0
torchvision>=0.22.0
torchaudio>=2.7.0
numpy>=1.26.0
Pillow>=10.0.0
gradio>=4.11.0
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-learn>=1.4.0
tqdm>=4.66.0
timm>=1.0.0